requirements.txt
Flask==3.0.3
tensorflow==2.16.1
Pillow==10.3.0
python-dotenv==1.0.1
google-cloud-storage==2.16.0
waitress==3.0.0